moved GoToFile to a separate file
