+++ src/OmniSharp/Api/Navigation/OmnisharpController.GotoDefinition.cs
@@ -7,19 +7,7 @@ using Microsoft.CodeAnalysis.Text;
 using OmniSharp.Models;
 
 namespace OmniSharp
-{
-    public partial class OmnisharpController
-    {
-        [HttpPost("gotofile")]
-        public async Task<QuickFixResponse> GoToFile(Request request)
-        {
-            var docs = _workspace.CurrentSolution.Projects.SelectMany(project => project.Documents).
-                Select(x => new QuickFix { FileName = x.FilePath, Text = x.Name, Line = 1, Column = 1});
-                
-            return new QuickFixResponse(docs);
-        }
-    }
-    
+{   
     public partial class OmnisharpController
     {
         [HttpPost("gotodefinition")]
